spanconfig: deflake spanconfigreconcilerccl/TestDataDriven
Fixes #98038. This test set up two protection records over two schema
objects at two timestamps ts=3 and ts=4.

   /Table/10{6-7}  protection_policies=[{ts: 3} {ts: 4}]
   /Table/10{7-8}  protection_policies=[{ts: 3} {ts: 4}]

When it later released those protection records:

   release record-id=3
   release record-id=4
   ----

It asserted that the span config mutations showed that we did infact get
rid of the protected state:

   mutations
   ----
   delete /Table/10{6-7}
   upsert /Table/10{6-7}      range default
   delete /Table/10{7-8}
   upsert /Table/10{7-8}      range default

But since release of these protections was non-atomic, in #98038 we
observed the following transition instead.

   delete /Table/10{6-7}
   upsert /Table/10{6-7}      protection_policies=[{ts: 4}]
   delete /Table/10{7-8}
   upsert /Table/10{7-8}      protection_policies=[{ts: 4}]
   delete /Table/10{6-7}
   upsert /Table/10{6-7}      range default
   delete /Table/10{7-8}
   upsert /Table/10{7-8}      range default

That is, we first got rid of the record with ts=3 and only then got rid
of ts=4. We just rewrite the test to assert on the final state of the
records that show no remaining protections, instead of trying to add
synchronization for mutations.

Release note: None